# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: DriveSaathi Vehicle Tracking and Management Server API
  description: DriveSaathi Vehicle Tracking and Management Server API documentation.
  contact:
    name: DriveSaathi Support
    url: https://www.drivesaathi.com/
    email: support@drivesaathi.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: "6.9.1"
servers:
  - url: http://51.21.203.186:8082/api
    description: Production Server
  - url: https://server.drivesaathi.com/api
    description: Subscription Server
  - url: http://localhost:8082/api
    description: Local Development (8082)
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /trips:
    get:
      summary: Fetch a list of Trips
      description: Without params, it returns a list of Trips the user has access to
      tags:
        - Trips
      parameters:
        - name: all
          in: query
          description: Can only be used by admins or managers to fetch all entities
          schema:
            type: boolean
        - name: userId
          in: query
          description: Standard users can use this only with their own _userId_
          schema:
            type: integer
        - name: deviceId
          in: query
          description: >-
            Standard users can use this only with _deviceId_s, they have access
            to
          schema:
            type: integer
        - name: groupId
          in: query
          description: >-
            Standard users can use this only with _groupId_s, they have access
            to
          schema:
            type: integer
        - name: refresh
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results to return (default 100)
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip for pagination (default 0)
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      summary: Create a Trip
      tags:
        - Trips
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
      x-codegen-request-body-name: body
  /trips/{id}:
    get:
      summary: Fetch a single Trip by ID
      tags:
        - Trips
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content: {}
    put:
      summary: Update a Trip
      tags:
        - Trips
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Trip"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
      x-codegen-request-body-name: body
    delete:
      summary: Delete a Trip
      tags:
        - Trips
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content
          content: {}
components:
  schemas:
...
    Trip:
      type: object
      required:
        - name
        - deviceId
        - originLatitude
        - originLongitude
        - destinationLatitude
        - destinationLongitude
        - etaDate
        - consignorId
        - consigneeId
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
          description: Trip name or identifier (optional)
        deviceId:
          type: integer
          format: int64
          description: "Device ID (internal identifier) - can provide either deviceId or deviceUniqueId"
        deviceUniqueId:
          type: string
          description: Device unique identifier (can be used instead of deviceId for creation)
        deviceName:
          type: string
          description: Device name (auto-filled)
          readOnly: true
        assetNumber:
          type: string
          description: Asset number (auto-filled from device uniqueId)
          readOnly: true
        assetType:
          type: string
          description: Type of asset (auto-filled from device attribute 'assetType')
          readOnly: true
        transporterName:
          type: string
          description: Name of the transporter (auto-filled from device's group name)
          readOnly: true
        otherTransporterName:
          type: string
          description: Other transporter name if custom entry needed
        consignorId:
          type: integer
          format: int64
          description: "Consignor user ID (required) - Use actual userId or a special value for 'other'"
        consignorName:
          type: string
          description: Name of the consignor (auto-filled from user if userId provided)
          readOnly: true
        otherConsignorName:
          type: string
          description: Custom consignor name (required if consignorId doesn't match an existing user)
        consigneeId:
          type: integer
          format: int64
          description: "Consignee user ID (required) - Use actual userId or a special value for 'other'"
        consigneeName:
          type: string
          description: Name of the consignee (auto-filled from user if userId provided)
          readOnly: true
        otherConsigneeName:
          type: string
          description: Custom consignee name (required if consigneeId doesn't match an existing user)
        consigneePincode:
          type: string
          description: Consignee postal code
        originLatitude:
          type: number
          format: double
          description: Origin latitude coordinate
        originLongitude:
          type: number
          format: double
          description: Origin longitude coordinate
        originAddress:
          type: string
          description: Origin address (optional)
        destinationLatitude:
          type: number
          format: double
          description: Destination latitude coordinate
        destinationLongitude:
          type: number
          format: double
          description: Destination longitude coordinate
        destinationAddress:
          type: string
          description: Destination address (optional)
        departureDate:
          type: string
          format: date-time
          description: Departure date/time (auto-set to current time)
          readOnly: true
        etaDate:
          type: string
          format: date-time
          description: Estimated time of arrival in ISO 8601 format (required, 24-hour format)
        bookingDate:
          type: string
          format: date-time
          description: Booking date and time in ISO 8601 format (optional)
        driverNumber:
          type: string
          description: "Driver unique number/ID for dropdown selection (optional) - backend will auto-fill driver details"
        driverName:
          type: string
          description: Driver name (auto-filled from driver record)
          readOnly: true
        driverMobile:
          type: string
          description: Driver mobile number (optional)
        driverContact:
          type: string
          description: Driver contact number (auto-filled from driver attributes)
          readOnly: true
        driverLicense:
          type: string
          description: Driver license number (auto-filled from driver attributes)
          readOnly: true
        driverId:
          type: integer
          format: int64
          description: Driver ID if linked to driver entity (optional)
        shipperName:
          type: string
          description: Shipper name (optional)
        shipmentId:
          type: string
          description: Unique shipment ID (optional)
        waybillNumber:
          type: string
          description: Waybill/LR number (optional)
        waybillExpiryDate:
          type: string
          format: date-time
          description: Waybill expiry date in ISO 8601 format (optional)
        waybillImage:
          type: string
          description: Path or URL to waybill image (optional)
        material:
          type: string
          description: Material type being transported (optional)
        otherMaterial:
          type: string
          description: Other material if "OTHERS" is selected (optional)
        invoiceNumber:
          type: string
          description: Invoice number (optional)
        invoiceQuantity:
          type: string
          description: Invoice quantity (e.g., in MT) (optional)
        lrGrNumber:
          type: string
          description: LR/GR number (optional)
        cardNumber:
          type: string
          description: Card number (optional)
        vehicleType:
          type: string
          description: Type of vehicle (optional)
        vehicleNumber:
          type: string
          description: Vehicle registration number (optional)
        currentLatitude:
          type: number
          format: double
          description: Current device latitude (transient - calculated from latest position)
          readOnly: true
        currentLongitude:
          type: number
          format: double
          description: Current device longitude (transient - calculated from latest position)
          readOnly: true
        currentSpeed:
          type: number
          format: double
          description: Current device speed (transient - from latest position)
          readOnly: true
        lastPositionUpdate:
          type: string
          format: date-time
          description: Last position update timestamp (transient)
          readOnly: true
        referencePositionId:
          type: integer
          format: int64
          description: Reference position ID (first position after departure)
        totalDistance:
          type: number
          format: double
          description: Total route distance in kilometers (origin to destination via ORS API)
          readOnly: true
        distanceCovered:
          type: number
          format: double
          description: Distance covered in kilometers (transient - origin to current position via ORS API)
          readOnly: true
        waypoint1:
          type: string
          description: First waypoint (optional)
        waypoint2:
          type: string
          description: Second waypoint (optional)
        runningStatus:
          type: string
          description: Running status of the trip (auto-set to 'pending' initially)
          readOnly: true
        dataDelay:
          type: string
          description: Data delay information (auto-updated)
          readOnly: true
        deviceType:
          type: string
          description: Type of device - Permanent, Temporary (optional)
        ownMarket:
          type: string
          description: Own or Market vehicle (optional)
        sda:
          type: string
          format: date-time
          description: SDA (Scheduled Departure Actual) date (optional)
        deta:
          type: string
          format: date-time
          description: DETA (Destination Estimated Time Actual) date (optional)
        createdAt:
          type: string
          format: date-time
          description: Trip creation timestamp (auto-generated by backend)
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Trip last update timestamp (auto-generated by backend)
          readOnly: true
        attributes:
          type: object
          properties: {}
          description: Custom attributes for the trip (optional)
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer